Move eval function from _Plot[] to mathics.eval...
Also clean up mathics.drawing.plot module
