+++ mathics/eval/plot.py
@@ -1,0 +1,385 @@
+"""
+Evaluation routines for 2D plotting.
+"""
+from math import cos, isinf, isnan, pi, sqrt
+from typing import Callable, Iterable, List, Optional, Union
+
+from mathics.builtin.numeric import chop
+from mathics.builtin.options import options_to_rules
+from mathics.builtin.scoping import dynamic_scoping
+from mathics.core.atoms import Real, String
+from mathics.core.convert.expression import to_mathics_list
+from mathics.core.convert.python import from_python
+from mathics.core.evaluation import Evaluation
+from mathics.core.expression import Expression
+from mathics.core.list import ListExpression
+from mathics.core.symbols import SymbolN, SymbolPower
+from mathics.core.systemsymbols import (
+    SymbolGraphics,
+    SymbolHue,
+    SymbolLine,
+    SymbolMessageName,
+    SymbolPoint,
+    SymbolQuiet,
+)
+
+RealPoint6 = Real(0.6)
+RealPoint2 = Real(0.2)
+
+
+try:
+    from mathics.compile import CompileArg, CompileError, _compile, real_type
+
+    has_compile = True
+except ImportError:
+    has_compile = False
+
+
+def automatic_plot_range(values):
+    """Calculates mean and standard deviation, throwing away all points
+    which are more than 'thresh' number of standard deviations away from
+    the mean. These are then used to find good vmin and vmax values. These
+    values can then be used to find Automatic Plotrange."""
+
+    if not values:
+        return 0, 1
+
+    thresh = 2.0
+    values = sorted(values)
+    valavg = sum(values) / len(values)
+    valdev = sqrt(
+        sum([(x - valavg) ** 2 for x in values]) / zero_to_one(len(values) - 1)
+    )
+
+    n1, n2 = 0, len(values) - 1
+    if valdev != 0:
+        for v in values:
+            if abs(v - valavg) / valdev < thresh:
+                break
+            n1 += 1
+        for v in values[::-1]:
+            if abs(v - valavg) / valdev < thresh:
+                break
+            n2 -= 1
+
+    vrange = values[n2] - values[n1]
+    vmin = values[n1] - 0.05 * vrange  # 5% extra looks nice
+    vmax = values[n2] + 0.05 * vrange
+    return vmin, vmax
+
+
+def compile_quiet_function(expr, arg_names, evaluation, list_is_expected: bool):
+    """
+    Given an expression return a quiet callable version.
+    Compiles the expression where possible.
+    """
+    if has_compile and not list_is_expected:
+        try:
+            cfunc = _compile(
+                expr, [CompileArg(arg_name, real_type) for arg_name in arg_names]
+            )
+        except CompileError:
+            pass
+        else:
+
+            def quiet_f(*args):
+                try:
+                    result = cfunc(*args)
+                    if not (isnan(result) or isinf(result)):
+                        return result
+                except Exception:
+                    pass
+                return None
+
+            return quiet_f
+    expr = Expression(SymbolN, expr).evaluate(evaluation)
+    quiet_expr = Expression(
+        SymbolQuiet,
+        expr,
+        ListExpression(Expression(SymbolMessageName, SymbolPower, String("infy"))),
+    )
+
+    def quiet_f(*args):
+        vars = {arg_name: Real(arg) for arg_name, arg in zip(arg_names, args)}
+        value = dynamic_scoping(quiet_expr.evaluate, vars, evaluation)
+        if list_is_expected:
+            if value.has_form("List", None):
+                value = [extract_pyreal(item) for item in value.elements]
+                if any(item is None for item in value):
+                    return None
+                return value
+            else:
+                return None
+        else:
+            value = extract_pyreal(value)
+            if value is None or isinf(value) or isnan(value):
+                return None
+            return value
+
+    return quiet_f
+
+
+def eval_Plot(
+    functions: List[Expression],
+    apply_fn: Callable,
+    x_name: str,
+    start: int,
+    stop: int,
+    x_range: list,
+    y_range,
+    plotpoints: int,
+    mesh,
+    list_is_expected: bool,
+    exclusions: list,
+    max_recursion: int,
+    options: dict,
+    evaluation: Evaluation,
+) -> Expression:
+    """
+    Evaluation part of Plot[]
+
+    Note: (?) indicates somewhat vague guesses.
+
+    functions: is a list of Mathics M-Expressions to be evaluated
+    start: minimum x-axis value
+    stop:  maximum t x-axis value
+    x_name; the name of the function parameter name used by ``functions``
+    x_range: x-axis range of the form Automatic, All, or [min, max]
+    y_range: y-axis range of the form Automatic, All, or [min, max]
+    y_range: either Automatic, All, or of the form [min, max]
+    plotpoints: number of points to plot
+    list_is_expected: list is expected in evaluation (?)
+    max_recursion: maximum number of levels of recursion in evaluation (?)
+    options: Plot options
+    evaluation: Expression evaluation object typically needed in evaluation
+    """
+    # constants to generate colors
+    hue = 0.67
+    hue_pos = 0.236068
+    hue_neg = -0.763932
+
+    def get_points_minmax(points):
+        xmin = xmax = ymin = ymax = None
+        for line in points:
+            for x, y in line:
+                if xmin is None or x < xmin:
+                    xmin = x
+                if xmax is None or x > xmax:
+                    xmax = x
+                if ymin is None or y < ymin:
+                    ymin = y
+                if ymax is None or y > ymax:
+                    ymax = y
+        return xmin, xmax, ymin, ymax
+
+    def get_points_range(points):
+        xmin, xmax, ymin, ymax = get_points_minmax(points)
+        if xmin is None or xmax is None:
+            xmin, xmax = 0, 1
+        if ymin is None or ymax is None:
+            ymin, ymax = 0, 1
+        return zero_to_one(xmax - xmin), zero_to_one(ymax - ymin)
+
+    function_hues = []
+    base_plot_points = []  # list of points in base subdivision
+    plot_points = []  # list of all plotted points
+    mesh_points = []
+    graphics = []  # list of resulting graphics primitives
+    for index, f in enumerate(functions):
+        points = []
+        xvalues = []  # x value for each point in points
+        tmp_mesh_points = []  # For this function only
+        continuous = False
+        d = (stop - start) / (plotpoints - 1)
+        compiled_fn = compile_quiet_function(f, [x_name], evaluation, list_is_expected)
+        for i in range(plotpoints):
+            x_value = start + i * d
+            point = apply_fn(compiled_fn, x_value)
+            if point is not None:
+                if continuous:
+                    points[-1].append(point)
+                    xvalues[-1].append(x_value)
+                else:
+                    points.append([point])
+                    xvalues.append([x_value])
+                continuous = True
+            else:
+                continuous = False
+
+        base_points = []
+        for line in points:
+            base_points.extend(line)
+        base_plot_points.extend(base_points)
+
+        xmin, xmax = automatic_plot_range([xx for xx, yy in base_points])
+        xscale = 1.0 / zero_to_one(xmax - xmin)
+        ymin, ymax = automatic_plot_range([yy for xx, yy in base_points])
+        yscale = 1.0 / zero_to_one(ymax - ymin)
+
+        if mesh == "System`Full":
+            for line in points:
+                tmp_mesh_points.extend(line)
+
+        def find_excl(excl):
+            # Find which line the exclusion is in
+            for line in range(len(xvalues)):  # TODO: Binary Search faster?
+                if xvalues[line][0] <= excl and xvalues[line][-1] >= excl:
+                    break
+                if (
+                    xvalues[line][-1] <= excl
+                    and xvalues[min(line + 1, len(xvalues) - 1)][0] >= excl
+                ):
+                    return min(line + 1, len(xvalues) - 1), 0, False
+            xi = 0
+            for xi in range(len(xvalues[line]) - 1):
+                if xvalues[line][xi] <= excl and xvalues[line][xi + 1] >= excl:
+                    return line, xi + 1, True
+            return line, xi + 1, False
+
+        if exclusions != "System`None":
+            for excl in exclusions:
+                if excl != "System`Automatic":
+                    l, xi, split_required = find_excl(excl)
+                    if split_required:
+                        xvalues.insert(l + 1, xvalues[l][xi:])
+                        xvalues[l] = xvalues[l][:xi]
+                        points.insert(l + 1, points[l][xi:])
+                        points[l] = points[l][:xi]
+                # assert(xvalues[l][-1] <= excl  <= xvalues[l+1][0])
+
+        # Adaptive Sampling - loop again and interpolate highly angled
+        # sections
+
+        # Cos of the maximum angle between successive line segments
+        ang_thresh = cos(pi / 180)
+
+        for line, line_xvalues in zip(points, xvalues):
+            recursion_count = 0
+            smooth = False
+            while not smooth and recursion_count < max_recursion:
+                recursion_count += 1
+                smooth = True
+                i = 2
+                while i < len(line):
+                    vec1 = (
+                        xscale * (line[i - 1][0] - line[i - 2][0]),
+                        yscale * (line[i - 1][1] - line[i - 2][1]),
+                    )
+                    vec2 = (
+                        xscale * (line[i][0] - line[i - 1][0]),
+                        yscale * (line[i][1] - line[i - 1][1]),
+                    )
+                    try:
+                        angle = (vec1[0] * vec2[0] + vec1[1] * vec2[1]) / sqrt(
+                            (vec1[0] ** 2 + vec1[1] ** 2)
+                            * (vec2[0] ** 2 + vec2[1] ** 2)
+                        )
+                    except ZeroDivisionError:
+                        angle = 0.0
+                    if abs(angle) < ang_thresh:
+                        smooth = False
+                        incr = 0
+
+                        x_value = 0.5 * (line_xvalues[i - 1] + line_xvalues[i])
+
+                        point = apply_fn(compiled_fn, x_value)
+                        if point is not None:
+                            line.insert(i, point)
+                            line_xvalues.insert(i, x_value)
+                            incr += 1
+
+                        x_value = 0.5 * (line_xvalues[i - 2] + line_xvalues[i - 1])
+                        point = apply_fn(compiled_fn, x_value)
+                        if point is not None:
+                            line.insert(i - 1, point)
+                            line_xvalues.insert(i - 1, x_value)
+                            incr += 1
+
+                        i += incr
+                    i += 1
+
+        if exclusions == "System`None":  # Join all the Lines
+            points = [[(xx, yy) for line in points for xx, yy in line]]
+
+        graphics.append(Expression(SymbolHue, Real(hue), RealPoint6, RealPoint6))
+        graphics.append(Expression(SymbolLine, from_python(points)))
+
+        for line in points:
+            plot_points.extend(line)
+
+        if mesh == "System`All":
+            for line in points:
+                tmp_mesh_points.extend(line)
+
+        if mesh != "System`None":
+            mesh_points.append(tmp_mesh_points)
+
+        function_hues.append(hue)
+
+        if index % 4 == 0:
+            hue += hue_pos
+        else:
+            hue += hue_neg
+        if hue > 1:
+            hue -= 1
+        if hue < 0:
+            hue += 1
+
+    x_range = get_plot_range(
+        [xx for xx, yy in base_plot_points], [xx for xx, yy in plot_points], x_range
+    )
+    y_range = get_plot_range(
+        [yy for xx, yy in base_plot_points], [yy for xx, yy in plot_points], y_range
+    )
+
+    options["System`PlotRange"] = from_python([x_range, y_range])
+
+    if mesh != "None":
+        for hue, points in zip(function_hues, mesh_points):
+            graphics.append(Expression(SymbolHue, Real(hue), RealPoint6, RealPoint6))
+            meshpoints = [to_mathics_list(xx, yy) for xx, yy in points]
+            graphics.append(Expression(SymbolPoint, ListExpression(*meshpoints)))
+
+    return Expression(
+        SymbolGraphics, ListExpression(*graphics), *options_to_rules(options)
+    )
+
+
+def extract_pyreal(value) -> Optional[float]:
+    if isinstance(value, Real):
+        return chop(value).round_to_float()
+    return None
+
+
+def get_plot_range(values: Iterable, all_values: Iterable, option: str) -> tuple:
+    """
+    Returns a tuple of the min and max values in values or
+    """
+    if option == "System`Automatic":
+        result = automatic_plot_range(values)
+    elif option == "System`All":
+        if not all_values:
+            result = (0, 1)
+        else:
+            result = min(all_values), max(all_values)
+    else:
+        result = option
+    if result[0] == result[1]:
+        value = result[0]
+        if value > 0:
+            return 0, value * 2
+        if value < 0:
+            return value * 2, 0
+        return -1, 1
+    return result
+
+
+def zero_to_one(value: Union[float, int]) -> Union[float, int]:
+    """
+    Return 1 only if ``value`` is zero, otherwise keep the value as is.
+
+    This is useful in scaling when the value can be used as
+    a divisor or when determining the number of points to plot and we want to
+    assure there is at least one point plotted.
+    """
+    return 1 if value == 0 else value
